added admin roles for create-update-delete rights

+++ services/equationServices.js
@@ -1,11 +1,16 @@
 const Equations = require('../models/Equations');
 const Topics = require('../models/Topics');
+const roleService = require('./roleService');
 
 async function getEquations(topicId) {
     return await Equations.find({topicId: topicId});
 }
 
-async function createEquation(equationParams) {
+async function createEquation(equationParams,uid) {
+    const role = await roleService.findUserRole(uid);
+    if (!role.isAdmin) {
+        throw `You Don't Have Admin Access`;
+    }
     const Topic = await Topics.findById(equationParams.topicId);
     if (!Topic) {
         throw `Topic for the Equation doesnot exist.`;
@@ -20,7 +25,11 @@ async function createEquation(equationParams) {
         equation: _savedEquation
     }
 }
-async function updateEquation(id, equationParams) {
+async function updateEquation(id, equationParams,uid) {
+    const role = await roleService.findUserRole(uid);
+    if (!role.isAdmin) {
+        throw `You Don't Have Admin Access`;
+    }
     const equation = await Equations.findById(id);
     if(!equation) throw `Equation not found.`;
     if (equationParams.topicId) {
@@ -37,7 +46,11 @@ async function updateEquation(id, equationParams) {
     }
 }
 
-async function deleteEquation(id) {
+async function deleteEquation(id,uid) {
+    const role = await roleService.findUserRole(uid);
+    if (!role.isAdmin) {
+        throw `You Don't Have Admin Access`;
+    }
     return await Equations.deleteOne({_id: id});
 }
 module.exports = {
